refactor(elixir): move rust native functions ockam_vault_software app
