+++ implementations/e/applications/ockam_vault_software/.gitignore
@@ -1,2 +1,3 @@
 _build
 deps
+cmake-build-*
