--- implementations/elixir/native/CMakeLists.txt
@@ -1,5 +1,0 @@
-cmake_minimum_required(VERSION 3.1...3.18 FATAL_ERROR)
-
-project(ockam_elixir VERSION 0.9.0 LANGUAGES C)
-
-add_subdirectory(vault)
