--- implementations/elixir/native/vault/CMakeLists.txt
@@ -1,10 +1,0 @@
-
-# add all subdirectories
-
-file(GLOB children RELATIVE ${CMAKE_CURRENT_SOURCE_DIR} ${CMAKE_CURRENT_SOURCE_DIR}/*)
-
-foreach(child ${children})
-  if(IS_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/${child})
-    add_subdirectory(${child})
-  endif()
-endforeach()
